pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend", "frontend", "shared", "infra"]

[project]
name = "golf-ui-streamlit"
version = "0.1.0"
description = "Python project to display personal golf analytics using streamlit"
authors = [
    { name = "Rhys powell"}
]

dependencies = [
    "azure-storage-blob==12.24.0",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
backend = [
    "selenium==4.25.0"
]
frontend = [
    "streamlit==1.46.0",
    "Authlib==1.5.1",
    "pandas==2.2.3",
    "plotly==5.24.1",
    "streamlit-components @ git+https://github.com/powellrhys/python-streamlit-components.git@v0.1.5"
]
testing = [
    "behave==1.2.6",
    "pytest==8.3.4",
    "pytest-cov==5.0.0"
]

[project.scripts]
collect-trackman-data = "backend.collect_trackman_data:main"
collect-scorecard-data = "backend.collect_scorecard_data:main"
